fix(gui): improve spk signature loading speed (fix #1827)
