+++ jadx-gui/src/main/resources/i18n/Messages_de_DE.properties
@@ -384,6 +384,7 @@ certificate.serialSHA256=SHA-256-Fingerabdruck
 certificate.serialPubKeyY=Y
 
 apkSignature.signer=Signierer
+#apkSignature.loading=Loading signature...
 apkSignature.verificationSuccess=Signaturprüfung erfolgreich abgeschlossen
 apkSignature.verificationFailed=Signaturprüfung fehlgeschlagen
 apkSignature.signatureSuccess=Gültige APK-Signatur v%d gefunden
