+++ jadx-gui/src/main/resources/i18n/Messages_es_ES.properties
@@ -384,6 +384,7 @@ certificate.serialSHA256=Huella SHA-256
 certificate.serialPubKeyY=Y
 
 #apkSignature.signer=Signer
+#apkSignature.loading=Loading signature...
 #apkSignature.verificationSuccess=Signature verification succeeded
 #apkSignature.verificationFailed=Signature verification failed
 #apkSignature.signatureSuccess=Valid APK signature v%d found
