+++ jadx-gui/src/main/resources/i18n/Messages_id_ID.properties
@@ -384,6 +384,7 @@ certificate.serialSHA256=Sidik Jari SHA-256
 certificate.serialPubKeyY=Y
 
 apkSignature.signer=Penandatangan
+#apkSignature.loading=Loading signature...
 apkSignature.verificationSuccess=Verifikasi tanda tangan berhasil
 apkSignature.verificationFailed=Verifikasi tanda tangan gagal
 apkSignature.signatureSuccess=Tanda tangan APK yang valid v%d ditemukan
