+++ jadx-gui/src/main/resources/i18n/Messages_ko_KR.properties
@@ -384,6 +384,7 @@ certificate.serialSHA256=SHA-256 지문
 certificate.serialPubKeyY=Y
 
 apkSignature.signer=서명자
+#apkSignature.loading=Loading signature...
 apkSignature.verificationSuccess=서명 검증 성공
 apkSignature.verificationFailed=서명 검증 실패
 apkSignature.signatureSuccess=유효한 APK 서명 v%d을(를) 찾았습니다.
