+++ jadx-gui/src/main/resources/i18n/Messages_pt_BR.properties
@@ -384,6 +384,7 @@ certificate.serialSHA256=Assinatura digital SHA-256
 certificate.serialPubKeyY=Y
 
 apkSignature.signer=Assinador
+#apkSignature.loading=Loading signature...
 apkSignature.verificationSuccess=Verificação de assinatura bem-sucedida
 apkSignature.verificationFailed=Verificação de assinatura falhou
 apkSignature.signatureSuccess=Assinatura válida de apk v%d encontrada
