+++ jadx-gui/src/main/resources/i18n/Messages_ru_RU.properties
@@ -384,6 +384,7 @@ certificate.serialSHA256=SHA-256 Fingerprint
 certificate.serialPubKeyY=Y
 
 apkSignature.signer=Signer
+#apkSignature.loading=Loading signature...
 apkSignature.verificationSuccess=Signature verification succeeded
 apkSignature.verificationFailed=Signature verification failed
 apkSignature.signatureSuccess=Valid APK signature v%d found
