+++ jadx-gui/src/main/resources/i18n/Messages_zh_CN.properties
@@ -384,6 +384,7 @@ certificate.serialSHA256=SHA-256 签名
 certificate.serialPubKeyY=Y
 
 apkSignature.signer=签名者
+#apkSignature.loading=Loading signature...
 apkSignature.verificationSuccess=签名验证成功
 apkSignature.verificationFailed=签名验证失败
 apkSignature.signatureSuccess=找到有效的 APK 签名 v%d
