+++ jadx-gui/src/main/resources/i18n/Messages_zh_TW.properties
@@ -384,6 +384,7 @@ certificate.serialSHA256=SHA-256 指紋
 certificate.serialPubKeyY=Y
 
 apkSignature.signer=簽署者
+#apkSignature.loading=Loading signature...
 apkSignature.verificationSuccess=簽名驗證成功
 apkSignature.verificationFailed=簽名驗證失敗
 apkSignature.signatureSuccess=找到可用的 APK 簽名 v%d
